% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cSeg.R
\name{cSeg}
\alias{cSeg}
\title{Recursive Segementation by Conditional Inference}
\usage{
cSeg(formula, data, maxsegs = Inf, maxdepth = 10L, minsplit = 20L,
  minbucket = 7L, ...)
}
\arguments{
\item{formula}{a symbolic description of the model to be fit.}

\item{data}{a data frame that contains the variables in the model.}

\item{maxsegs}{maximal number of segments}

\item{maxdepth}{maximal depth of the tree models used for recursive segmentation. The number of decision rules that define a segment can be controled this way.}

\item{minsplit}{minimal size of a subset to allow for furhter segmentation.}

\item{minbucket}{minimal size of a segment.}

\item{...}{further arguments passed to \link[partykit]{ctree_control}.}
}
\value{
An object of class \code{segmentation}.
}
\description{
Recursive segmentation based on conditional inference trees.
}
\details{
The algorithm makes use of \link[partykit]{ctree} to construct a recursive segmentation model. There is a \link[=predict.segmentation]{predict}, \link[=predict.segmentation]{plot}, \link[=predict.segmentation]{summary} and \link[=print.segmentation]{print} function. The . See the corresponding documentation for details.
}
\examples{
### regression
airq <- subset(airquality, !is.na(Ozone))
airseg <- cSeg(Ozone ~ ., data = airq)
airseg
plot(airseg)

# Here, observations with missing values are processed by surrogate splits
airseg2 <- cSeg(Ozone ~ ., data = airq, maxsurrogate = 1)
airseg2
plot(airseg2)

### classification
irisseg <- cSeg(Species ~ ., data = iris)
irisseg
plot(irisseg)

### survival analysis
if (require("TH.data")) {
data("GBSG2", package = "TH.data")
GBSG2seg <- cSeg(Surv(time, cens) ~ ., data = GBSG2)
print(GBSG2seg)
plot(GBSG2seg)
}

### multivariate responses
airseg3 <- cSeg(Ozone + Temp ~ ., data = airq)
airseg3
}
\references{
{
\insertRef{Hapfelmeier2018}{rseg}
}
}
