% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eseg.R
\name{eseg}
\alias{eseg}
\title{Recursive Segmentation by Evolutionary Learning}
\usage{
eseg(formula, data, maxsegs = Inf, maxdepth = 10L, minsplit = 20L,
  minbucket = 7L, ...)
}
\arguments{
\item{formula}{a symbolic description of the model to be fit.}

\item{data}{a data frame that contains the variables in the model.}

\item{maxsegs}{maximal number of segments}

\item{maxdepth}{maximal depth of the tree models used for recursive segmentation. The number of decision rules that define a segment can be controled this way.}

\item{minsplit}{minimal size of a subset to allow for furhter segmentation.}

\item{minbucket}{minimal size of a segment.}

\item{...}{further arguments passed to \link[evtree]{evtree.control}.}
}
\value{
An object of class \code{rseg}.
}
\description{
Recursive segmentation based on evolutionary trees.
}
\details{
The algorithm makes use of \link[evtree]{evtree} to construct a recursive segmentation model. There is a \link[=predict.rseg]{predict}, \link[=plot.rseg]{plot}, \link[=summary.rseg]{summary} and \link[=print.rseg]{print} function. The \link[=gettree]{gettree} function can be used to extract the correspoding tree model. See the corresponding documentation for details. The function \link[evtree]{evtree} does currently not allow missing values in the covariates.
}
\examples{
### regression
airq <- subset(airquality, complete.cases(airquality))
set.seed(1234)
airseg <- eseg(Ozone ~ ., data = airq)
airseg
plot(airseg)

### classification
set.seed(1234)
irisseg <- eseg(Species ~ ., data = iris)
irisseg
plot(irisseg)
}
\references{
{
\insertRef{Hapfelmeier2018}{rseg}
}
}
