% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.rseg}
\alias{plot.rseg}
\title{Plot objects of class \code{rseg}}
\usage{
\method{plot}{rseg}(x, ...)
}
\arguments{
\item{x}{an object of class \code{rseg} fit by \link{cseg}, \link{eseg} or \link{rseg}.}

\item{...}{additional arguments passed to \link[partykit]{plot.constparty}.}
}
\value{
Plot of a \code{rseg} object.
}
\description{
\code{plot} method for objects of class \code{rseg}.
}
\details{
Does currently not support plotting of multivariate outcomes and censored outcomes for \link{eseg} and \link{rseg}.
}
\examples{
airq <- subset(airquality, !is.na(Ozone))
airct <- cseg(Ozone ~ ., data = airq)
airct
plot(airct)
}
