% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.segmentation}
\alias{plot.segmentation}
\title{Plot objects of class \code{segmentation}}
\usage{
\method{plot}{segmentation}(x, ...)
}
\arguments{
\item{x}{an object of class \code{segmentation} fit by \link{cSeg}, \link{eSeg} or \link{rSeg}.}

\item{...}{additional arguments passed to \link[partykit]{plot.constparty}.}
}
\value{
Plot of a \code{segmentation} object.
}
\description{
\code{plot} method for objects of class \code{segmentation}.
}
\details{
Does currently not support plotting of multivariate outcomes and censored outcomes for \link{eSeg} and \link{rSeg}.
}
\examples{
airq <- subset(airquality, !is.na(Ozone))
airct <- cSeg(Ozone ~ ., data = airq)
airct
plot(airct)
}
