% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.segmentation}
\alias{predict.segmentation}
\title{Predictions by recursive segmentation}
\usage{
\method{predict}{segmentation}(object, newdata = NULL, type = c("response",
  "prob", "segment")[1], ...)
}
\arguments{
\item{object}{an object of class \code{segmentation} fit by \link{cSeg}, \link{eSeg} or \link{rSeg}.}

\item{newdata}{optional data to use for pediction. If omitted, the fitted values in \code{object} are used.}

\item{type}{the type of prediction to be returned. For "response", this is the mean of a numeric response, the predicted class for a categorical response or the median survival time for a censored response. For "prob", there are two options: 1) in case of a categorical response this is a matrix of conditional class probabilities or a list with the conditional class probabilities for each observation. 2) In case of numeric and censored responses, this is a list with the empirical cumulative distribution functions and empirical survivor functions (Kaplan-Meier estimate). For "segment", this is an integer vector of terminal node identifiers.}

\item{...}{not used.}
}
\value{
A list, matrix or vector of predictions.
}
\description{
\code{predict} method for objects of class \code{segmentation}.
}
\examples{
### survival analysis
if (require("TH.data")) {
data("GBSG2", package = "TH.data")
GBSG2seg <- cSeg(Surv(time, cens) ~ ., data = GBSG2)
print(GBSG2seg)
predict(GBSG2seg, type = "response")
predict(GBSG2seg, type = "prob")
predict(GBSG2seg, type = "segment")
}
}
