% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.rseg}
\alias{print.rseg}
\title{Print objects of class \code{rseg}}
\usage{
\method{print}{rseg}(x, ...)
}
\arguments{
\item{x}{an object of class \code{rseg} fit by \link{cseg}, \link{eseg} or \link{rseg}.}

\item{...}{not used.}
}
\description{
\code{print} method for objects of class \code{rseg}.
}
\examples{
airq <- subset(airquality, !is.na(Ozone))
airct <- cseg(Ozone ~ ., data = airq)
airct
}
