% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.segmentation}
\alias{print.segmentation}
\title{Print objects of class \code{segmentation}}
\usage{
\method{print}{segmentation}(x, ...)
}
\arguments{
\item{x}{an object of class \code{segmentation} fit by \link{cSeg}, \link{eSeg} or \link{rSeg}.}

\item{...}{not used.}
}
\description{
\code{print} method for objects of class \code{segmentation}.
}
\examples{
airq <- subset(airquality, !is.na(Ozone))
airct <- cSeg(Ozone ~ ., data = airq)
airct
}

