% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SegMOB.R
\name{SegMOB}
\alias{SegMOB}
\title{Model-based Recursive Segementation}
\usage{
SegMOB(formula, data, fit, maxsegs = Inf, maxdepth = 10L, minsplit = 20L,
  minbucket = 7L, ...)
}
\arguments{
\item{formula}{a symbolic description of the model to be fit.}

\item{data}{a data frame that contains the variables in the model.}

\item{maxsegs}{maximal number of segments}

\item{maxdepth}{maximal depth of the tree models used for recursive segmentation. The number of decision rules that define a segment can be controled this way.}

\item{minsplit}{minimal size of a subset to allow for furhter segmentation.}

\item{minbucket}{minimal size of a segment.}

\item{...}{further arguments passed to \link[partykit]{mob_control}.}
}
\value{
An object of class \code{segmentation}.
}
\description{
Recursive segmentation based on Model-based Recursive Partitioning.
}
\details{
The algorithm makes use of \link[partykit]{mob} to construct a model-based recursive segmentation model. There is a \link[=predict.segmentation]{predict}, \link[=plot.segmentation]{plot}, \link[=summary.segmentation]{summary} and \link[=print.segmentation]{print} function. The \link[=gettree]{gettree} function can be used to extract the correspoding tree model. See the corresponding documentation for details.
}
\examples{
if(require("mlbench")) {
  
  ## Pima Indians diabetes data
  data("PimaIndiansDiabetes", package = "mlbench")
  
  ## a simple basic fitting function (of type 1) for a logistic regression
  logit <- function(y, x, start = NULL, weights = NULL, offset = NULL, ...) {
    glm(y ~ 0 + x, family = binomial, start = start, ...)
  }
  
  ## set up a logistic regression tree
  pid_tree <- SegMOB(diabetes ~ glucose | pregnant + pressure + triceps + insulin +
                    mass + pedigree + age, data = PimaIndiansDiabetes, fit = logit)
  
  ## print tree
  print(pid_tree)
  
  ## visualization
  plot(pid_tree)
 
  ## summary
  summary(pid_tree)
  
  ## predicted nodes
  predict(pid_tree, newdata = head(PimaIndiansDiabetes, 6), type = "segment")
}
}
\references{
{
\insertRef{Hapfelmeier2018}{rseg}
}
}
