% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.rseg}
\alias{summary.rseg}
\title{Summary of objects of class \code{rseg}}
\usage{
\method{summary}{rseg}(object, ...)
}
\arguments{
\item{object}{an object of class \code{rseg} fit by \link{cseg}, \link{eseg} or \link{rseg}.}

\item{...}{not used.}
}
\value{
A list of three to four elements: 1) \code{segments} contains the number of segments in \code{object}, \code{rules} contains a vector of the decision rules that define the segments in \code{object} and 3) \code{size} contains a vector of segment sizes in \code{object}. 4) In case of Model-based Recursive Segementation there is a fourth element containing the summaries of the models fit to the segments.
}
\description{
\code{summary} method for objects of class \code{rseg}.
}
\details{
This function provides some information about \code{rseg} objects.
}
\examples{
airq <- subset(airquality, !is.na(Ozone))
airct <- cseg(Ozone ~ ., data = airq)
summary(airct)
}
