% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.segmentation}
\alias{summary.segmentation}
\title{Summary of objects of class \code{segmentation}}
\usage{
\method{summary}{segmentation}(object, ...)
}
\arguments{
\item{object}{an object of class \code{segmentation} fit by \link{cSeg}, \link{eSeg} or \link{rSeg}.}

\item{...}{not used.}
}
\value{
A list of three elements: 1) \code{segments} contains the number of segments in \code{object}, \code{rules} contains a vector of the decision rules that define the segments in \code{object} and 3) \code{size} contains a vector of segment sizes in \code{object}.
}
\description{
\code{summary} method for objects of class \code{segmentation}.
}
\details{
This function provides some information about \code{segmentation} objects.
}
\examples{
airq <- subset(airquality, !is.na(Ozone))
airct <- cSeg(Ozone ~ ., data = airq)
summary(airct)
}
