% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gettree.R
\name{gettree}
\alias{gettree}
\title{Extract tree model}
\usage{
gettree(x)
}
\arguments{
\item{x}{an object of class \code{rseg} fit by \link{cseg}, \link{eseg} or \link{rseg}.}
}
\value{
A tree model fit by the functions \link[partykit]{ctree}, \link[evtree]{evtree} or \link[rpart]{rpart}.
}
\description{
Extract a tree model from an object of class \code{rseg}.
}
\details{
Recursive segmentation by \link{cseg}, \link{eseg} or \link{rseg} fits models by iterative application of the functions \link[partykit]{ctree}, \link[evtree]{evtree} or \link[rpart]{rpart}, respectively. Therefore, it is possible to extract a respective tree model from an object of class \code{rseg}.
}
\examples{
airq <- subset(airquality, !is.na(Ozone))
airct <- cseg(Ozone ~ ., data = airq)
aircttree <- gettree(airct)
aircttree
plot(aircttree)
}
