% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmob.R
\name{segmob}
\alias{segmob}
\title{Model-based Recursive Segementation}
\usage{
segmob(formula, data, fit, weights = NULL, maxsegs = Inf,
  maxdepth = 10L, minsplit = 20L, minbucket = 7L, ...)
}
\arguments{
\item{formula}{a symbolic description of the model to be fit.}

\item{data}{a data frame that contains the variables in the model.}

\item{fit}{A function for fitting the model within each node. For details see \link[partykit]{mob}.}

\item{weights}{optional variable of weights contained in data. By default these are treated as case weights but the default can be changed in \link[partykit]{mob_control}.}

\item{maxsegs}{maximal number of segments}

\item{maxdepth}{maximal depth of the tree models used for recursive segmentation. The number of decision rules that define a segment can be controled this way.}

\item{minsplit}{minimal size of a subset to allow for further segmentation.}

\item{minbucket}{minimal size of a segment.}

\item{...}{further arguments passed to \link[partykit]{mob_control}.}
}
\value{
An object of class \code{rseg}.
}
\description{
Recursive segmentation based on Model-based Recursive Partitioning.
}
\details{
The algorithm makes use of \link[partykit]{mob} to construct a model-based recursive segmentation model. There is a \link[=predict.rseg]{predict}, \link[=plot.rseg]{plot}, \link[=summary.rseg]{summary} and \link[=print.rseg]{print} function. The \link[=gettree]{gettree} function can be used to extract the correspoding tree model. See the corresponding documentation for details.
}
\examples{
if(require("AER")) {
  
  data("TeachingRatings", package = "AER")
  tr <- subset(TeachingRatings, credits == "more")
  
  segmob(eval ~ beauty | minority + age + gender + division + native + tenure, 
    data = tr, fit = partykit:::lmfit, weights = students, caseweights = FALSE)
}
}
\references{
{
\insertRef{Hapfelmeier2018}{rseg}
}
}
